% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_implausibility_alcohol_mean.R
\name{calculate_implausibility_alcohol_mean}
\alias{calculate_implausibility_alcohol_mean}
\title{function to calculate implausibility for alcohol continuous calibration
this outputs a TP}
\usage{
calculate_implausibility_alcohol_mean(data, brfss)
}
\description{
function to calculate implausibility for alcohol continuous calibration
this outputs a TP
}
\examples{
calculate_implausibility_alcohol_mean
}
\keyword{implausibility}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
