% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_implausibility_alcohol.R,
%   R/calculate_implausibility_education.R
\name{calculate_implausibility_education}
\alias{calculate_implausibility_education}
\title{function to calculate implausibility for education transitions calibration
this outputs a TP}
\usage{
calculate_implausibility_education(data, targets)

calculate_implausibility_education(data, targets)
}
\description{
function to calculate implausibility for education transitions calibration
this outputs a TP

function to calculate implausibility for education transitions calibration
this outputs a TP
}
\examples{
calculate_implausibility_education
calculate_implausibility_education
}
\keyword{implausibility}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
