% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_implausibility_mortality.R
\name{calculate_implausibility_mortality}
\alias{calculate_implausibility_mortality}
\title{function to calculate implausibility for mortality calibration
this outputs a TP}
\usage{
calculate_implausibility_mortality(
  data,
  agest = 1,
  agestyear = 2010,
  model_error
)
}
\description{
function to calculate implausibility for mortality calibration
this outputs a TP
}
\examples{
calculate_implausibility_mortality
}
\keyword{implausibility}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
