% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_oneyear.R
\name{convert_to_one_year}
\alias{convert_to_one_year}
\title{function to convert 3 yr NESARC TPs to 1 yr TPs
this outputs a TP}
\usage{
convert_to_one_year(probs)
}
\description{
function to convert 3 yr NESARC TPs to 1 yr TPs
this outputs a TP
}
\examples{
convert_to_one_year
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
