% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_probability.R
\name{convert_to_probability}
\alias{convert_to_probability}
\title{function to convert new sampled markov paramters to transition probabilities
this outputs a TP}
\usage{
convert_to_probability(samples, model, covariates)
}
\description{
function to convert new sampled markov paramters to transition probabilities
this outputs a TP
}
\examples{
sample_from_markov
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
