% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcov2numeric.msm.R
\name{factorcov2numeric.msm}
\alias{factorcov2numeric.msm}
\title{function to convert the covariates to the correct format for markov model calibration}
\usage{
factorcov2numeric.msm(covariates, x, mod = NULL)
}
\description{
function to convert the covariates to the correct format for markov model calibration
}
\examples{
convert covariates to correct format for markov model
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
