% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_targets_alcohol.R
\name{generate_targets_alcohol}
\alias{generate_targets_alcohol}
\title{function to generate alcohol use calibration targets}
\usage{
generate_targets_alcohol(data)
}
\description{
function to generate alcohol use calibration targets
}
\examples{
generate_targets_alcohol
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
\keyword{targest}
