% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation_wrapper.R
\name{optimisation_wrapper}
\alias{optimisation_wrapper}
\title{function to run optimisation process running microsimulation and extracting uncertainty}
\usage{
optimisation_wrapper(x, model, targets, variance, maxyear, threshold)
}
\description{
function to run optimisation process running microsimulation and extracting uncertainty
}
\keyword{code}
\keyword{microsimulation}
\keyword{optimisation}
\keyword{wrapper}
