% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_markov_model.R
\name{resample_markov_model}
\alias{resample_markov_model}
\title{function for resampling the markov model based on restricted parameter space}
\usage{
resample_markov_model(newsamples)
}
\description{
function for resampling the markov model based on restricted parameter space
}
\examples{
resample_markov_model
}
\keyword{implausibility}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
