% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_mortality_lhs.R
\name{resample_mortality_lhs}
\alias{resample_mortality_lhs}
\title{function for resampling the risk function parameters for mortality}
\usage{
resample_mortality_lhs(models, topsamples, nsamples)
}
\description{
function for resampling the risk function parameters for mortality
}
\examples{
resample_mortality_lhs
}
\keyword{implausibility}
\keyword{microsimulation}
\keyword{model}
\keyword{multinomial}
