% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_markov_model.R,
%   R/resample_multinom_model.R
\name{resample_multinom_model}
\alias{resample_multinom_model}
\title{function for resampling the multinomial logistic regression model for alcohol}
\usage{
resample_multinom_model(models, topsamples, nsamples)

resample_multinom_model(models, topsamples, nsamples)
}
\description{
function for resampling the multinomial logistic regression model for alcohol

function for resampling the multinomial logistic regression model for alcohol
}
\examples{
resample_multinom_model
resample_multinom_model
}
\keyword{implausibility}
\keyword{microsimulation}
\keyword{model}
\keyword{multinomial}
