% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_ordinal_model.R
\name{resample_ordinal_model}
\alias{resample_ordinal_model}
\title{function for resampling the multinomial logistic regression model for alcohol}
\usage{
resample_ordinal_model(models, topsamples, nsamples)
}
\description{
function for resampling the multinomial logistic regression model for alcohol
}
\examples{
resample_ordinal_model
}
\keyword{implausibility}
\keyword{microsimulation}
\keyword{model}
\keyword{multinomial}
