% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_markov.R
\name{sample_from_markov}
\alias{sample_from_markov}
\title{function to convert the covariates to the correct format for markov model calibration}
\usage{
sample_from_markov(model, nsamples, inflation, originalsample, inflatedsample)
}
\description{
function to convert the covariates to the correct format for markov model calibration
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
