% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_markov_inflated.R
\name{sample_from_markov_inflated}
\alias{sample_from_markov_inflated}
\title{function to convert the covariates to the correct format for markov model calibration}
\usage{
sample_from_markov_inflated(
  model,
  nsamples,
  inflation,
  originalsample,
  inflatedsample,
  category
)
}
\description{
function to convert the covariates to the correct format for markov model calibration
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
