% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Q.R
\name{setup_Q}
\alias{setup_Q}
\title{function to setup the Q transition matrix for markov models for calibration}
\usage{
setup_Q(model, estimates)
}
\description{
function to setup the Q transition matrix for markov models for calibration
}
\examples{
setup_Q
}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
