% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model_alcohol_output.R
\name{summarise_model_alcohol_output}
\alias{summarise_model_alcohol_output}
\title{function to summarise alcohol use output from microsimulation}
\usage{
summarise_model_alcohol_output(data)
}
\description{
function to summarise alcohol use output from microsimulation
}
\examples{
summarise_alcohol_output
}
\keyword{microsimulation}
